requirements.txt
# Django (Legacy - removal pending if unneeded)
Django>=5.0
django-cors-headers

# Supabase
supabase>=2.0.0

# Streamlit
streamlit>=1.30.0
streamlit-authenticator>=0.2.0

# Database
psycopg2-binary>=2.9

# Environment variables
python-dotenv>=1.0.0

# Data Visualization & Other utilities
Pillow>=10.0.0
plotly>=5.18.0
pandas>=2.0.0

# AI APIs
groq>=0.4.0
requests>=2.31.0
